pyproject.toml
[project]
name = "dlp"
version = "0.1.0"
description = ""
authors = [
    {name = "Oleg",email = "wondertopp@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.119.0,<0.120.0)",
    "uvicorn[standart] (>=0.38.0,<0.39.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pydantic-settings (>=2.11.0,<3.0.0)",
    "websockets (>=15.0.1,<16.0.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "aiosqlite (>=0.21.0,<0.22.0)",
    "flashtext (>=2.7,<3.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
